requirement.txt
acres==0.3.0
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
beautifulsoup4==4.13.3
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.1.8
configobj==5.0.9
configparser==7.2.0
distro==1.9.0
Django==4.2.20
etelemetry==0.3.1
exceptiongroup==1.2.2
faiss-cpu==1.10.0
filelock==3.18.0
fitz==0.0.1.dev2
frontend==0.0.3
google==3.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
importlib_resources==6.5.2
isodate==0.6.1
itsdangerous==2.2.0
jiter==0.9.0
looseversion==1.3.0
lxml==5.3.1
networkx==3.2.1
nibabel==5.3.2
nipype==1.10.0
numpy==2.0.2
openai==1.70.0
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
proto-plus==1.26.1
protobuf==5.29.4
prov==2.0.1
puremagic==1.28
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.1
pydantic_core==2.33.0
pydot==3.0.4
PyMuPDF==1.25.5
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytils==0.4.1
pytz==2025.2
pyxnat==1.6.3
rdflib==6.3.2
requests==2.32.3
rsa==4.9
scipy==1.13.1
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.3
starlette==0.46.1
tools==0.1.9
tqdm==4.67.1
traits==7.0.2
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
zipp==3.21.0